created auth middleware for server
