+++ auth-test/auth-middleware.js
@@ -1,0 +1,91 @@
+
+var db = require('../database.js');
+var users = db.get('users')
+
+var session = require('express-session');
+var passport = require('passport');
+var OAuth2Strategy = require('passport-oauth').OAuth2Strategy;
+var GitHubStrategy = require('passport-github').Strategy;
+
+module.exports = function(app){
+
+  app.get('/',function(req, res, next){
+    console.log('hello')
+    if (req.user || req.session.passport !== {}) {
+      console.log('in the checkUser utility function, about to execute next()');
+      next();
+    } else {
+      console.log('in the checkUser utility function, about to have the response object send "0"');
+      res.send('0');
+    }
+  });
+
+  passport.serializeUser(function(user, done){
+    console.log('user about to be serialized', user);
+    done(null, user);
+  });
+
+  passport.deserializeUser(function(user, done) {
+    done(null, user);
+  });
+
+  app.use(session({
+    saveUninitialized: true, 
+    resave: true, 
+    secret: 'this is our secret'
+  }));
+
+  // Authentication Strategy for Github
+
+  passport.use(new GitHubStrategy({
+    clientID: 'your client id',
+    clientSecret: 'your client secret',
+    callbackURL: 'http://localhost:3000/auth/github/callback'
+
+    // Begin Authentication Callback 
+
+  }, function(accessToken, refreshToken, profile, done){
+
+      // Logging for Testing
+
+      console.log('accessToken', accessToken);
+      console.log('profile', profile);
+
+      // Find Document with matching userid and then save access token
+
+      // users.findOne({userid: profile.id}).on('success', function (doc) {
+      //   doc.accessToken = accessToken;
+      //   return done(null, profile);
+      // });
+
+      // For Testing
+
+      process.nextTick(function () {
+         return done(null, profile);
+       });
+    }
+  ));
+
+  app.use(passport.initialize());
+  app.use(passport.session());
+
+  app.get('/auth/github',
+    passport.authenticate('github', {scope: 'write:repo_hook'} ));
+
+  app.get('/auth/github/callback', 
+    passport.authenticate('github', 
+      { sucessRedirect: '/', failureRedirect: '/login' }),
+    function(req, res) {
+      res.redirect('/');
+    });
+
+  app.get('/auth/error', function(req, res){
+    console.log('login failed in server');
+    res.redirect('/');
+  });
+
+  app.get('/loggedin', function(req, res){
+    res.send(req.isAuthenticated() ? req.user : 'unauthorized');
+  });
+
+}
