+++ auth-test/test-server.js
@@ -1,7 +1,7 @@
 'use strict';
 
-var db = require('../database.js');
-var users = db.get('users')
+// var db = require('../database.js');
+// var users = db.get('users')
 
 // User Collection Schema
 // {userid: INT, accessToken: STRING }
@@ -12,89 +12,89 @@ var bodyParser = require('body-parser');
 var cors = require('cors');
 var app = express();
 
-var session = require('express-session');
-var passport = require('passport');
-var OAuth2Strategy = require('passport-oauth').OAuth2Strategy;
-var GitHubStrategy = require('passport-github').Strategy;
+// var session = require('express-session');
+// var passport = require('passport');
+// var OAuth2Strategy = require('passport-oauth').OAuth2Strategy;
+// var GitHubStrategy = require('passport-github').Strategy;
 
 app.use(cors());
 
 app.use(express.static(__dirname + '/client/'));
 
-app.get('/',function(req, res, next){
-  console.log('hello')
-  if (req.user || req.session.passport !== {}) {
-    console.log('in the checkUser utility function, about to execute next()');
-    next();
-  } else {
-    console.log('in the checkUser utility function, about to have the response object send "0"');
-    res.send('0');
-  }
-});
+require('./auth-middleware.js')(app)
 
-passport.serializeUser(function(user, done){
-  console.log('user about to be serialized', user);
-  done(null, user);
-});
+// app.get('/',function(req, res, next){
+//   console.log('hello')
+//   if (req.user || req.session.passport !== {}) {
+//     console.log('in the checkUser utility function, about to execute next()');
+//     next();
+//   } else {
+//     console.log('in the checkUser utility function, about to have the response object send "0"');
+//     res.send('0');
+//   }
+// });
 
-passport.deserializeUser(function(user, done) {
-  done(null, user);
-});
+// passport.serializeUser(function(user, done){
+//   console.log('user about to be serialized', user);
+//   done(null, user);
+// });
 
-app.use(session({
-  saveUninitialized: true, 
-  resave: true, 
-  secret: 'this is our secret'
-}));
+// passport.deserializeUser(function(user, done) {
+//   done(null, user);
+// });
 
-// Authentication Strategy for Github
+// app.use(session({
+//   saveUninitialized: true, 
+//   resave: true, 
+//   secret: 'this is our secret'
+// }));
 
-passport.use(new GitHubStrategy({
-  clientID: 'secret',
-  clientSecret: 'secret',
-  callbackURL: 'http://localhost:3000/auth/github/callback'
+// // Authentication Strategy for Github
 
-  // Begin Authentication Callback 
+// passport.use(new GitHubStrategy({
+//   clientID: 'secret',
+//   clientSecret: 'secret',
+//   callbackURL: 'http://localhost:3000/auth/github/callback'
 
-}, function(accessToken, refreshToken, profile, done){
+//   // Begin Authentication Callback 
 
-    // Logging for Testing
+// }, function(accessToken, refreshToken, profile, done){
 
-    console.log('accessToken', accessToken);
-    console.log('profile', profile);
+//     // Logging for Testing
 
-    // Find Document with matching userid and then save access token
+//     console.log('accessToken', accessToken);
+//     console.log('profile', profile);
 
-    users.findOne({userid: profile.id}).on('success', function (doc) {
-      doc.accessToken = accessToken;
-      return done(null, profile);
-    });
-  }
-));
+//     // Find Document with matching userid and then save access token
 
-users.findOne({ name: 'test' }).on('success', function (doc) {});
+//     users.findOne({userid: profile.id}).on('success', function (doc) {
+//       doc.accessToken = accessToken;
+//       return done(null, profile);
+//     });
+//   }
+// ));
 
-app.use(passport.initialize());
-app.use(passport.session());
+// app.use(passport.initialize());
+// app.use(passport.session());
 
-app.get('/auth/github',
-  passport.authenticate('github', {scope: 'write:repo_hook'} ));
+// app.get('/auth/github',
+//   passport.authenticate('github', {scope: 'write:repo_hook'} ));
 
-app.get('/auth/github/callback', 
-  passport.authenticate('github', 
-    { sucessRedirect: '/', failureRedirect: '/login' }),
-  function(req, res) {
-    res.redirect('/');
-  });
+// app.get('/auth/github/callback', 
+//   passport.authenticate('github', 
+//     { sucessRedirect: '/', failureRedirect: '/login' }),
+//   function(req, res) {
+//     res.redirect('/');
+//   });
 
-app.get('/auth/error', function(req, res){
-  console.log('login failed in server');
-  res.redirect('/');
-});
+// app.get('/auth/error', function(req, res){
+//   console.log('login failed in server');
+//   res.redirect('/');
+// });
 
-app.get('/loggedin', function(req, res){
-  res.send(req.isAuthenticated() ? req.user : 'unauthorized');
-});
+// app.get('/loggedin', function(req, res){
+//   res.send(req.isAuthenticated() ? req.user : 'unauthorized');
+// });
 
 var server = app.listen(3000, function(){
   var host = server.address().address;
